feat: improve flows on account change (#93)
* feat: improve flows on account change

When signing in, signing up or logging out, invalidate the 'decks' key
to ensure the sidebar stays up to date.

Also redirects /register to /review on sign-up. Could be a smarter
redirect in the future that checks if you have cards to review and
otherwise sends you to a potential /home page or triggers a deck
creation flow.

* redirect to /review instead of /decks on sign up